UPD: Reviews

+++ backend/reviews/models.py
@@ -8,10 +8,8 @@ from django.core.validators import MinValueValidator, MaxValueValidator
 class Review(models.Model):
     user = models.ForeignKey(CustomUser, on_delete=models.CASCADE)
     text = models.TextField()
-    rating = models.DecimalField(
-        max_digits=3,
-        decimal_places=2,
-        validators=[MinValueValidator(0), MaxValueValidator(5)],
+    rating = models.PositiveIntegerField(
+        validators=[MinValueValidator(0), MaxValueValidator(5)]
     )
     created_at = models.DateTimeField(auto_now_add=True)
